# shared/specs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Edge AI Platform API
  version: 1.0.0
  description: |
    REST API for the Edge AI Platform web management interface.
    Provides device management, chat, and metrics access.
  contact:
    name: VLMChat Platform
  license:
    name: MIT
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: http://localhost:3000
    description: Docker development server
security:
  - bearerAuth: []
...
paths:
...
  /api/config/device:
    get:
      tags: [devices]
      summary: Get device type configuration
      description: Retrieve the active configuration for a device type (e.g., macos, raspberry-pi, jetson)
      operationId: getDeviceConfig
      parameters:
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [macos, raspberry-pi, jetson, generic, other]
          description: Device type/platform
      responses:
        '200':
          description: Device configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceConfig'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [devices]
      summary: Update device type configuration
      description: Update configuration for all devices of this type
      operationId: updateDeviceConfig
      parameters:
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [macos, raspberry-pi, jetson, generic, other]
          description: Device type/platform
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [config]
              properties:
                config:
                  type: object
                  description: Full device configuration (tasks and sinks)
                changeDescription:
                  type: string
                  description: Optional description of what changed
      responses:
        '200':
          description: Configuration updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceConfig'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/config/device/history:
    get:
      tags: [devices]
      summary: Get device type configuration history
      description: Retrieve configuration change history for a device type
      operationId: getDeviceConfigHistory
      parameters:
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [macos, raspberry-pi, jetson, generic, other]
          description: Device type/platform
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 100
      responses:
        '200':
          description: Configuration history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConfigHistory'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/config/device/rollback/{historyId}:
    post:
      tags: [devices]
      summary: Rollback device type configuration
      description: Restore a previous configuration version for a device type
      operationId: rollbackDeviceConfig
      parameters:
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [macos, raspberry-pi, jetson, generic, other]
          description: Device type/platform
        - name: historyId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Configuration rolled back
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceConfig'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/config/framework:
    get:
      tags: [devices]
      summary: Get framework configuration
      description: Retrieve shared pipeline configuration
      operationId: getFrameworkConfig
      responses:
        '200':
          description: Framework configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkConfig'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [devices]
      summary: Update framework configuration
      description: Update shared pipeline configuration
      operationId: updateFrameworkConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [config]
              properties:
                name:
                  type: string
                  default: default
                config:
                  type: object
                  description: Framework configuration (pipeline, sources, buffers)
                changeDescription:
                  type: string
      responses:
        '200':
          description: Configuration updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrameworkConfig'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/config/framework/history:
    get:
      tags: [devices]
      summary: Get framework configuration history
      description: Retrieve framework configuration change history
      operationId: getFrameworkConfigHistory
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 100
      responses:
        '200':
          description: Configuration history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConfigHistory'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
...
  schemas:
...
    DeviceConfig:
      description: Database record for device configuration with version tracking
      allOf:
        - $ref: '../schemas/device-config-v1.0.0.json'
    FrameworkConfig:
      description: Database record for framework configuration with version tracking
      allOf:
        - $ref: '../schemas/framework-config-v1.0.0.json'
    ConfigHistory:
      description: Audit record for configuration changes
      allOf:
        - $ref: '../schemas/config-history-v1.0.0.json'
  responses:
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: string
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Access denied"
              message:
                type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: JWT session token from cookie-based authentication